pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "finetune-erp"
version = "0.1.0"
description = "Finetune ERP monorepo - repair shop management system"
readme = "README.md"
requires-python = ">=3.11"
dependencies = []
authors = [
    {name = "Finetune ERP Team"}
]
keywords = ["erp", "repair-shop", "django", "react", "monorepo"]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

[tool.setuptools]
packages = []
py-modules = []

[project.optional-dependencies]
backend = [
    "Django==5.2.5",
    "dj-database-url==2.3.0",
    "django-cors-headers==4.7.0",
    "djangorestframework==3.16.1",
    "djangorestframework-simplejwt==5.5.1",
    "gunicorn==23.0.0",
    "python-dotenv==1.1.1",
    "requests==2.32.5",
    "twilio==7.7.1",
    "weasyprint==66.0",
    "whitenoise==6.9.0",
    "psycopg2-binary>=2.9",
    "django-csp==4.0",
]
dev = []
test = [
    "pytest==8.4.1",
    "pytest-django==4.11.1",
    "freezegun==1.5.5",
    "black==24.4.2",
]
docs = [
    "sphinx>=7.0",
    "sphinx-rtd-theme>=2.0",
    "sphinx-autodoc-typehints>=1.25",
]

[tool.black]
line-length = 88
